admin/admin-panel-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.16.0
python-dotenv==1.0.0
pydantic==2.9.2
PyJWT==2.8.0
python-multipart==0.0.6
cloudinary==1.36.0